requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.10
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==3.0
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.0.8
django-crispy-forms==1.9.2
djangorestframework==3.11.1
dlib==19.21.1
Faker==4.1.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
Markdown==3.2.2
numpy==1.20.1
oauthlib==3.1.0
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
Pillow==7.2.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.10
whitenoise==5.2.0
wincertstore==0.2
zipp==3.1.0
zope.interface==5.3.0
